backend/pipeline/requirements.txt
pypdf==4.2.0
pdfminer.six==20231228
